linting and serializer clean up. Use getOptions not getSettings.
