+++ packages/dashboard/src/app/serializers/templates.js
@@ -17,7 +17,7 @@
 /**
  * External dependencies
  */
-import { toUTCDate, toDate, getSettings } from '@web-stories-wp/date';
+import { toUTCDate, toDate, getOptions } from '@web-stories-wp/date';
 /**
  * Internal dependencies
  */
@@ -50,8 +50,9 @@ export default function reshapeTemplateObject(
     slug,
     postersByPage,
     centerTargetAction: `${APP_ROUTES.TEMPLATE_DETAIL}?id=${id}&isLocal=${isLocal}`,
-    creationDate: toDate(creationDate, getSettings()),
+    creationDate: toDate(creationDate, getOptions()),
     status: 'template',
     modified: toUTCDate(modified),
+    isLocal,
   };
 }
